SimpsonsFunctions/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
onnxruntime
pillow!=8.3.0
requests